+++ NEWS.md
@@ -8,3 +8,11 @@
 
 * `irepeat()`: Iterator that returns an object indefinitely unless a specified
   number of times is given.
+
+* `iseq()`: Iterator for sequence generation. Follows the base `seq()` function.
+
+* `iseq_along()`: Iterators for sequence generation of length equal to an
+  object. Follows the base `seq_along()` function.
+
+* `iseq_len()`: Iterators for sequence generation of a fixed length. Follows the
+  base `seq_len()` function.